Binance/requirements.txt
pip==20.1.1
setuptools==47.1.0
python-binance
TA-Lib
numpy
websocket_client